add EdgeDetails and fix a deprecated
