+++ interactive_engine/executor/ir/graph_proxy/src/adapters/gs_store/mod.rs
@@ -13,6 +13,7 @@
 //! See the License for the specific language governing permissions and
 //! limitations under the License.
 
+mod details;
 mod partitioner;
 mod read_graph;
 mod translation;
